client: replace custom timeout with Set{Write,Read}Deadline
Based on quick benchmark, the current timeout implementation turned out
to have a quite significant impact due to the spawning of additional go
routines and the memory allocations caused by time.After. Let's switch
to configuring read and write deadlines instead. On average, on my laptop,
with a 0 dispatch interval and a single replica, the number of iterations
per second increased from ~35k to ~50k.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>