Connectivity disruption test overhaul
This commit makes many improvements to both the client and server that would
be tricky to split up into smaller commits, as the program was relatively compact
and the changes themselves aren't very layered. As a result, this has basically
turned into a rewrite of the client. Here's a small overview:

- Made the writer and reader asynchronous to allow for accurate packet scheduling.
  Before, writing and reading happened in a single loop, which puts a hard limit
  on the packet rate since latency is the dominating factor. (unless sleep() is
  removed)
- Message size reduced to 16 bytes since we're more interested in testing pps than
  throughput. This cuts out a significant part of the CPU consumption, which was
  starting to weigh on Cilium CI where this stack is run.
- Dynamic sleep interval based on socket write duration + scheduler jitter. The
  client will automatically try to attain the requested packet interval, enqueueing
  packets as evenly-spaced as possible. This was completely broken in the initial
  implementation, since latency wasn't accounted for. See below for example output.
- Added a --latency flag to directly control the read timeout for replies.
- dispatch-interval=0 no longer busy-polls and is automatically rewritten to 500us
  for backwards compatibility purposes. Existing Cilium tests currently specify 0ms
  to send packets as fast as possible, but with the new implementation, this will
  quickly fill the tx buffer and give the tcp stack a small meltdown.
- Client set itself to nice19 to avoid stealing too much CPU time from the Cilium
  agent in resource-constrained environments. (aka. 4-core LVH VMs with 3-node kind
  clusters)
- Writer is now pinned to an OS thread to reduce scheduling jitter caused by the Go
  runtime.
- nanosleep(2) is called directly from the writer goroutine to avoid the Go
  runtime's sleep coalescing and throttling. (see `strace -fenanosleep` on a typical
  Go program calling time.Sleep() in a tight loop)
- The client now does a best effort to exit gracefully to avoid connection resets
  being logged on the server, which otherwise pollutes the log.

- Moved most of the code out of main().
- Sets itself to nice19 like the client.
- Close all sockets before terminating.

On the server side, I experimented with graceful termination/socket draining,
but since the client drives message production, sending a signal to the client
to terminate would involve introducing a higher-level protocol. If the server
Pod is shut down, that typically means the whole test namespace is being killed,
so clean logs aren't much of a priority in that case.

Some example output of the client:

```
~ go run ./cmd/client localhost:1234 --dispatch-interval 0
Zero interval changed to 500µs for backwards compatibility, otherwise bufferbloat will interfere.
Connected to [::1]:1234 from [::1]:34620
Sending requests at a target interval of 500µs with max expected latency of 250ms
Operations per second: tx 2001, rx 2001, 31.3KiB/s
Operations per second: tx 2001, rx 2001, 31.3KiB/s
^CReader shutting down
Writer shutting down
```

Signed-off-by: Timo Beckers <timo@isovalent.com>